RPC server launches successfully
Can use threading to make it asynchronus

Using msgpack as a method of transport to make data handling easier
also allows us to keep data sizes small instead of using something like
json to do what we need. Will need to make sure this is the only
dependency. As it should be lightweight as possible

So far it is behaving as expected but will need to keep an eye on it
when working with the client side code.